Move to Spree 3.6 and Rails 5.2

+++ db/migrate/20170201153018_remove_environment_from_tracker.spree.rb
@@ -1,7 +1,9 @@
 # This migration comes from spree (originally 20150128032538)
 class RemoveEnvironmentFromTracker < ActiveRecord::Migration[4.2]
   def up
-    Spree::Tracker.where('environment != ?', Rails.env).update_all(active: false)
-    remove_column :spree_trackers, :environment
+    if defined?(Spree::Tracker)
+      Spree::Tracker.where('environment != ?', Rails.env).update_all(active: false)
+      remove_column :spree_trackers, :environment
+    end
   end
 end
